refactor(control): split shutdown orchestration
